소스 정리, 구조 정리
