--- src/main/java/com/become/atum/webportal/common/exception/RollbackException.java
@@ -1,13 +1,0 @@
-package com.become.atum.webportal.common.exception;
-
-public class RollbackException extends Exception {
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1558926996963203363L;
-
-	public RollbackException(String message) {
-		super(message);
-	}
-}
