+++ src/main/java/com/project/webproject/common/exception/RollbackException.java
@@ -1,0 +1,13 @@
+package com.project.webproject.common.exception;
+
+public class RollbackException extends Exception {
+
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1558926996963203363L;
+
+	public RollbackException(String message) {
+		super(message);
+	}
+}
