+++ src/main/java/com/project/webproject/model/UserGroup.java
@@ -1,0 +1,13 @@
+package com.project.webproject.model;
+
+import java.io.Serializable;
+
+import lombok.Data;
+
+@Data
+public class UserGroup implements Serializable {
+
+    private String userGroupId;
+    private String userGroupName;
+
+}
