+++ src/main/java/com/project/webproject/service/CommonService.java
@@ -1,0 +1,5 @@
+package com.project.webproject.service;
+
+public class CommonService {
+    
+}
