+++ src/main/java/com/project/webproject/service/MemberService.java
@@ -1,0 +1,5 @@
+package com.project.webproject.service;
+
+public class MemberService {
+    
+}
