+++ src/main/resources/application.properties
@@ -8,12 +8,10 @@ version = v5.0.1
 
 mariadb.atum.staserver.host = 192.168.0.2:8071
 
-prj2.staserver.host = 192.168.0.27:8888
-
 khepri.database.type       = mariadb
-khepri.database.ip         = 192.168.0.28
-khepri.database.username   = ATUM_DEPLOY
-khepri.database.password   = ATUM_DEPLOY
+khepri.database.ip         = localhost
+khepri.database.username   = DEVELOPMENT
+khepri.database.password   = DEVELOPMENT
 
 spring.datasource.hikari.maximum-pool-size=10
 #spring.datasource.hikari.connection-test-query=SELECT 1 FROM SYS.DUAL
@@ -32,9 +30,6 @@ mybatis.configuration.jdbc-type-for-null=NULL
 # mybatis.mapper-locations=mybatis/**/*.xml
 
 
-
-
-
 #logging config
 logging.config=classpath:logback-spring.xml
 
@@ -50,8 +45,5 @@ spring.devtools.livereload.enabled=true
 
 ##attach file path
 attach.file.path=d:\attachFiles
-
-##UX-Designer 시뮬레이션모드
-atum.simulationMode=N
 
 ##JAVA 경로
