+++ src/main/webapp/resources/js/functions/Board/main-Rule.js
