+++ src/main/webapp/resources/js/functions/Board/ruleModeling-Rule.js
